requirements.txt
paramiko
python-dotenv
cryptography